+++ src/components/TouchBar/Popover.js
@@ -1,7 +1,7 @@
 import { remote } from 'electron';
 import { forwardRef, useImperativeHandle, useLayoutEffect, useRef } from 'react';
 
-import { useElementRefValue } from './useElementRefValue';
+import { useElementRefValue } from '../../hooks/useElementRefValue';
 
 export const Popover = forwardRef(function Popover({
 	children: itemsElement,
