+++ src/hooks/useElementsRefValues.js
@@ -8,11 +8,21 @@ const isEachElementTheSame = (a, b) => {
 	return Array.from({ length: a.length }).every((_, i) => Object.is(a[i], b[i]));
 };
 
+const toArray = (elements) => [].concat(...Children.toArray(elements).map((element) => {
+	if (!element.type.render) {
+		return toArray(element.props.children);
+	}
+
+	return element;
+}));
+
 export const useElementsRefValues = (elements) => {
+	elements = toArray(elements);
+
 	const [values, setValues] = useState([]);
-	const refs = useRef(Array.from({ length: Children.count(elements) }, () => createRef()));
-	refs.current = Array.from({ length: Children.count(elements) }, (_, i) => refs.current[i]);
-	const clonedElements = Children.map(elements, (element, i) => !!element && cloneElement(element, { ref: refs.current[i] }));
+	const refs = useRef(Array.from({ length: elements.length }, () => createRef()));
+	refs.current = Array.from({ length: elements.length }, (_, i) => refs.current[i]);
+	const clonedElements = elements.map((element, i) => !!element && cloneElement(element, { key: i, ref: refs.current[i] }));
 
 	// eslint-disable-next-line react-hooks/exhaustive-deps
 	useLayoutEffect(() => {
